Music generation/AMI/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=pyha770c72_0
bidict=0.22.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
boto3=1.24.2=py38h06a4308_0
botocore=1.27.2=py38h06a4308_0
brotlipy=0.7.0=py38h0a891b7_1004
ca-certificates=2022.4.26=h06a4308_0
certifi=2022.6.15=py38h06a4308_0
cffi=1.15.0=py38hd667e15_1
chardet=5.0.0=pypi_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
click=8.0.4=py38h06a4308_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=35.0.0=py38h3e25421_2
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.11.0=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
filelock=3.7.1=pyhd8ed1ab_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.34.4=pypi_0
freetype=2.11.0=h70c0345_0
gcc=5.4.0=0
gdown=4.5.1=pyhd8ed1ab_0
giflib=5.2.1=h7b6447c_0
glib=2.56.2=hd408876_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=he6710b0_3
idna=3.3=pyhd8ed1ab_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
ipdb=0.13.9=pyhd8ed1ab_0
ipykernel=6.9.1=py38h06a4308_0
ipython=8.3.0=py38h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py38h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
jmespath=0.10.0=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jsonpickle=2.2.0=pypi_0
jsonschema=4.4.0=py38h06a4308_0
jupyter=1.0.0=py38_7
jupyter_client=7.2.2=py38h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.37=hbc83047_0
libprotobuf=3.15.8=h780b84a_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=h2818925_1
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.5.2=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
midisynth=0.3=pypi_0
miditoolkit=0.1.16=pypi_0
mido=1.2.10=pypi_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
more-itertools=8.13.0=pypi_0
music21=7.3.3=pypi_0
muspy=0.5.0=pypi_0
nbclient=0.5.13=py38h06a4308_0
nbconvert=6.4.4=py38h06a4308_0
nbformat=5.3.0=py38h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py38h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nltk=3.7=pyhd3eb1b0_0
notebook=6.4.11=py38h06a4308_0
numpy=1.22.3=py38he7a7128_0
numpy-base=1.22.3=py38hf524024_0
openssl=1.1.1o=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py38hace64e9_1
pip=22.1.2=py38h06a4308_0
pretty-midi=0.2.9=pypi_0
progress=1.6=pyhd8ed1ab_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.15.8=py38h709712a_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd8ed1ab_0
pyfluidsynth=1.3.1=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pypianoroll=1.0.4=pypi_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=py38h578d9bd_5
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.7.0=py3.8_cuda10.1.243_cudnn7.6.3_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py38h295c915_2
qt=5.9.7=h5867ecd_1
qtconsole=5.3.0=pyhd3eb1b0_0
qtpy=2.0.1=pyhd3eb1b0_0
readline=8.1.2=h7f8727e_1
regex=2022.7.9=py38h5eee18b_0
requests=2.28.1=py38h06a4308_0
s3transfer=0.6.0=py38h06a4308_0
scipy=1.9.0=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py38h06a4308_0
sip=6.5.1=py38h295c915_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.38.5=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
stanfordcorenlp=3.9.1.1=pypi_0
tensorboardx=2.5.1=pyhd8ed1ab_0
terminado=0.13.1=py38h06a4308_0
testpath=0.6.0=py38h06a4308_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.7.0=py38
torchvision=0.8.0=py38_cu101
tornado=6.1=py38h27cfd23_0
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webcolors=1.12=pypi_0
webencodings=0.5.1=py38_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py38h06a4308_0
xz=5.2.5=h7f8727e_1
zeromq=4.3.4=h2531618_0
zipp=3.8.0=py38h06a4308_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
